requirements.txt
mysql-connector-python==8.1.0
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0
PyYAML==6.0.1
pytest==7.4.2
